+++ src/test/java/nextstep/subway/acceptance/LineSteps.java
@@ -3,12 +3,14 @@ package nextstep.subway.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import nextstep.common.Steps;
+
 import org.springframework.http.MediaType;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class LineSteps {
+public class LineSteps extends Steps {
     public static ExtractableResponse<Response> 지하철_노선_생성_요청(String name, String color) {
         Map<String, String> params = new HashMap<>();
         params.put("name", name);
