+++ src/test/java/nextstep/subway/acceptance/PathAcceptanceTest.java
@@ -3,6 +3,7 @@ package nextstep.subway.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import nextstep.common.AcceptanceTest;
 import nextstep.subway.domain.WeightType;
 
 import org.junit.jupiter.api.BeforeEach;
