+++ src/test/java/nextstep/subway/acceptance/StationSteps.java
@@ -3,12 +3,14 @@ package nextstep.subway.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import nextstep.common.Steps;
+
 import org.springframework.http.MediaType;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class StationSteps {
+public class StationSteps extends Steps {
     public static ExtractableResponse<Response> 지하철역_생성_요청(String name) {
         Map<String, String> params = new HashMap<>();
         params.put("name", name);
